dev/FLASK/requirements.txt
aiofile==3.8.8
aiofiles==23.2.1
amazon-transcribe==0.6.2
awscrt==0.16.26
bidict==0.22.1
blinker==1.6.3
boto3==1.28.71
botocore==1.31.71
caio==0.9.13
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
eventlet==0.33.3
Flask==3.0.0
Flask-SocketIO==5.3.6
gevent==23.9.1
greenlet==3.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.4
hyperframe==6.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
priority==2.0.0
PyAudio==0.2.13
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
Quart==0.19.3
requests==2.31.0
s3transfer==0.7.0
simple-websocket==1.0.0
six==1.16.0
SpeechRecognition==3.10.0
urllib3==2.0.7
uuid==1.30
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
